+++ tests/Frameworks/Symfony/Version_2_8/src/AppBundle/Controller/DefaultController.php
@@ -11,7 +11,7 @@ class DefaultController extends Controller
     /**
      * @Route("/", name="homepage")
      */
-    public function indexAction(Request $request)
+    public function testingRouteNameAction(Request $request)
     {
         // replace this example code with whatever you need
         return $this->render('default/index.html.twig', array(